Retain unknown attributes on Sentence Save
Following PerseusDL/perseids_docs#31 we want to be able to support
additional markup/attributes in treebank files, even if the editor
doesn't support curation of these.

Before this commit a new sentence was built from scratch by using
values right out of the editor session, now we update the attributes we
know of in place of the old document.