Remove all *-local.* files and getlist-upload.xq
It appears that all these files are out of use - discovered while
working on #2.
All of them seem to be replaced by files without the local suffix. They
were called from treebank-getlist-upload.xq which also looks like it's
not used anymore.

Revert this commit if this assumption was actually wrong.

There might be other files in the repo that are also outdated.